requirements.txt
advent-of-code-data==1.1.1
alabaster==0.7.12
astroid==2.9.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
docutils==0.18.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.0
isort==5.10.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.5
packaging==21.3
Pebble==4.6.3
platformdirs==2.4.0
Pygments==2.10.0
pylint==2.12.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
scipy==1.7.3
six==1.16.0
SnakeMD==0.9.3
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
sphinx-issues==1.2.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
termcolor==1.1.0
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wrapt==1.13.3
zipp==3.6.0

pytest>=6.2.5